Skeleton project structure
